Se desarrolló lo principal para  permitir a un usuario registrar una compra (con su detalle) en el sistema. Restan aún pulir algunos detalles

+++ database/migrations/2022_10_05_231055_create_purchases_table.php
@@ -33,9 +33,9 @@ return new class extends Migration
 
             $table->integer('voucher_number')->required()->unique();
 
-            $table->decimal('subtotal', 10, 2)->required();
-            $table->decimal('iva', 10, 2)->required();
-            $table->decimal('total', 10, 2)->required();
+            $table->float('subtotal')->required();
+            $table->float('iva')->required();
+            $table->float('total')->required();
 
             $table->decimal('weight', 10, 2)->required();
             $table->string('weight_voucher')->nullable();
